Move decode to utils inside frontity package
