+++ packages/frontity/src/utils/decode/server.ts
@@ -1,5 +1,5 @@
 import he from "he";
-import { Decode } from "../../types";
+import { Decode } from "./types";
 import simpleDecode from "simple-entity-decode";
 import containsHTMLEntities from "./containsHTMLEntities";
