+++ packages/frontity/src/utils/decode/types.ts
@@ -1,0 +1,3 @@
+export interface Decode {
+  (text: string): string;
+}
